requirements.txt
python-telegram-bot
selenium
webdriver-manager
requests